+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Ensure URLs are properly encoded in `NetHttp#download` and `#open` using Addressable (@linyaoli)
+
 * Raise `ResponseError` with clear message when redirect URI was invalid in Down::NetHttp (@janko-m)
 
 ## 4.5.0 (2018-05-11)